requirements.txt
flask
flask-cors
google-generativeai>=0.3.0  # Assurez-vous d'avoir une version récente
gunicorn